+++ Libplanet.Tests/Store/Trie/KeyValueStoreTest.cs
@@ -19,7 +19,7 @@ namespace Libplanet.Tests.Store.Trie
 
         protected Random Random { get; } = new Random();
 
-        private byte[][] PreStoredDataKeys { get; set; }
+        private KeyBytes[] PreStoredDataKeys { get; set; }
 
         private byte[][] PreStoredDataValues { get; set; }
 
@@ -38,9 +38,25 @@ namespace Libplanet.Tests.Store.Trie
         }
 
         [SkippableFact]
+        public void GetMany()
+        {
+            KeyBytes[] nonExistentKeys = Enumerable.Range(0, 10)
+                .Select(_ => NewRandomKey())
+                .ToArray();
+            KeyBytes[] keys = PreStoredDataKeys
+                .Concat(PreStoredDataKeys.Take(PreStoredDataCount / 2))
+                .Concat(nonExistentKeys)
+                .ToArray();
+            IReadOnlyDictionary<KeyBytes, byte[]> result = KeyValueStore.Get(keys);
+            Assert.Equal(PreStoredDataCount, result.Count);
+            Assert.All(PreStoredDataKeys, k => Assert.Contains(k, result));
+            Assert.All(nonExistentKeys, k => Assert.DoesNotContain(k, result));
+        }
+
+        [SkippableFact]
         public void Set()
         {
-            byte[] key = Random.NextBytes(PreStoredDataKeySize);
+            var key = new KeyBytes(Random.NextBytes(PreStoredDataKeySize));
             byte[] value = Random.NextBytes(PreStoredDataValueSize);
             KeyValueStore.Set(key, value);
 
@@ -50,16 +66,16 @@ namespace Libplanet.Tests.Store.Trie
         [SkippableFact]
         public void SetMany()
         {
-            var values = new Dictionary<byte[], byte[]>();
+            var values = new Dictionary<KeyBytes, byte[]>();
             foreach (int i in Enumerable.Range(0, 10))
             {
-                values[Random.NextBytes(PreStoredDataKeySize)] =
+                values[new KeyBytes(Random.NextBytes(PreStoredDataKeySize))] =
                     Random.NextBytes(PreStoredDataValueSize);
             }
 
             KeyValueStore.Set(values);
 
-            foreach (KeyValuePair<byte[], byte[]> kv in values)
+            foreach (KeyValuePair<KeyBytes, byte[]> kv in values)
             {
                 Assert.Equal(kv.Value, KeyValueStore.Get(kv.Key));
             }
@@ -86,12 +102,29 @@ namespace Libplanet.Tests.Store.Trie
         [SkippableFact]
         public void Delete()
         {
-            foreach (var (key, expectedValue) in PreStoredDataKeys.Zip(
-                PreStoredDataValues, ValueTuple.Create))
+            foreach (KeyBytes key in PreStoredDataKeys)
             {
-                var actual = KeyValueStore.Get(key);
-                Assert.Equal(expectedValue, actual);
+                KeyValueStore.Delete(key);
+                Assert.False(KeyValueStore.Exists(key));
             }
+
+            KeyBytes nonExistent = NewRandomKey();
+            KeyValueStore.Delete(nonExistent);
+            Assert.False(KeyValueStore.Exists(nonExistent));
+        }
+
+        [SkippableFact]
+        public void DeleteMany()
+        {
+            KeyBytes[] nonExistentKeys = Enumerable.Range(0, 10)
+                .Select(_ => NewRandomKey())
+                .ToArray();
+            KeyBytes[] keys = PreStoredDataKeys
+                .Concat(PreStoredDataKeys.Take(PreStoredDataCount / 2))
+                .Concat(nonExistentKeys)
+                .ToArray();
+            KeyValueStore.Delete(keys);
+            Assert.All(keys, k => Assert.False(KeyValueStore.Exists(k)));
         }
 
         [SkippableFact]
@@ -109,18 +142,17 @@ namespace Libplanet.Tests.Store.Trie
         [SkippableFact]
         public void ListKeys()
         {
-            ImmutableHashSet<byte[]> keys = KeyValueStore.ListKeys().ToImmutableHashSet();
+            ImmutableHashSet<KeyBytes> keys = KeyValueStore.ListKeys().ToImmutableHashSet();
             Assert.Equal(PreStoredDataCount, keys.Count);
-            var equalityComparer = new BytesEqualityComparer();
-            Assert.True(PreStoredDataKeys.ToImmutableHashSet(equalityComparer).SetEquals(keys));
+            Assert.True(PreStoredDataKeys.ToImmutableHashSet().SetEquals(keys));
         }
 
-        public byte[] NewRandomKey()
+        public KeyBytes NewRandomKey()
         {
-            byte[] randomKey;
+            KeyBytes randomKey;
             do
             {
-                randomKey = Random.NextBytes(PreStoredDataKeySize);
+                randomKey = new KeyBytes(Random.NextBytes(PreStoredDataKeySize));
             }
             while (KeyValueStore.Exists(randomKey));
 
@@ -129,12 +161,12 @@ namespace Libplanet.Tests.Store.Trie
 
         protected void InitializePreStoredData()
         {
-            PreStoredDataKeys = new byte[PreStoredDataCount][];
+            PreStoredDataKeys = new KeyBytes[PreStoredDataCount];
             PreStoredDataValues = new byte[PreStoredDataCount][];
 
             for (int i = 0; i < PreStoredDataCount; ++i)
             {
-                PreStoredDataKeys[i] = Random.NextBytes(PreStoredDataKeySize);
+                PreStoredDataKeys[i] = new KeyBytes(Random.NextBytes(PreStoredDataKeySize));
                 PreStoredDataValues[i] = Random.NextBytes(PreStoredDataValueSize);
                 KeyValueStore.Set(PreStoredDataKeys[i], PreStoredDataValues[i]);
             }