+++ Libplanet.Tests/TestUtils.cs
@@ -18,6 +18,7 @@ using Libplanet.Blocks;
 using Libplanet.Crypto;
 using Libplanet.Net.Protocols;
 using Libplanet.Store;
+using Libplanet.Store.Trie;
 using Libplanet.Tx;
 using Xunit;
 using Xunit.Abstractions;
@@ -221,6 +222,12 @@ Actual (C# array lit):   new byte[{actual.LongLength}] {{ {actualRepr} }}";
             AssertBytesEqual(expected.ToByteArray(), actual.ToByteArray());
 
         public static void AssertBytesEqual(Address? expected, Address? actual) =>
+            AssertBytesEqual(expected?.ToByteArray(), actual?.ToByteArray());
+
+        public static void AssertBytesEqual(KeyBytes expected, KeyBytes actual) =>
+            AssertBytesEqual(expected.ToByteArray(), actual.ToByteArray());
+
+        public static void AssertBytesEqual(KeyBytes? expected, KeyBytes? actual) =>
             AssertBytesEqual(expected?.ToByteArray(), actual?.ToByteArray());
 
         public static void AssertBencodexEqual(IValue expected, IValue actual)